Accelerate multi-qubit gates (#490)
* M  pennylane_lightning/core/src/bindings/Bindings.hpp; hack `JacobianData` to work with devices.
M  pennylane_lightning/core/src/simulators/lightning_kokkos/StateVectorKokkos.hpp; `applyMatrix` bugfix: use intermediate hostview to copy matrix data; same bugfix for `getDataVector`.
M  pennylane_lightning/core/src/simulators/lightning_kokkos/algorithms/AdjointJacobianKokkos.hpp; use copy constructor.
M  pennylane_lightning/core/src/simulators/lightning_kokkos/measurements/MeasurementsKokkos.hpp; use copy constructor.
M  pennylane_lightning/core/src/simulators/lightning_kokkos/observables/ObservablesKokkos.hpp; use copy constructor.
M  requirements-dev.txt; add clang-format-14.

* Auto update version

* Update changelog.

* Auto update version

* Auto update version

* Add an argument to adjointJacobian to avoid syncing and copying state vector data in adjoint-diff.

* Reformat

* trigger CI

* [skip ci] Update changelog.

* Introduce std::unordered_map<std::string, ExpValFunc> expval_funcs_.

* Introduce applyExpectationValueFunctor.

* Add binding to LKokkos expval(matrix, wires). Combine expval functor calls into two templated methods. Call specialized expval methods when possible. Remove obsolete 'Apply directly' tests.

* Update changelog.

* Add test for arbitrary expval(Hermitian).

* Add getExpectationValueMultiQubitOpFunctor.

* Add typename hint for macos.

* Add typename macos.

* Use Kokkos::ThreadVectorRange policy for innerloop in getExpectationValueMultiQubitOpFunctor.

* Auto update version

* Auto update version

* Couple fix for HIP.

* WIP

* Add specialized 3-5 qubit expval functors.

* Add implementation notes.

* [skip ci] Polish getExpValMatrix and limit getExpValNQubitOpFunctor to N == 4 (seems optimal on OPENMP/HIP backends although CUDA can go further).

* Fix whitespace warning.

* Auto update version

* WIP

* Add new NQubit functors + bugfix

* Use simplified bitshifts in 1- & 2-qubit expval kernels.

* Do not template NQubit methods over inverse and just take conj-trans before calling the functor.

* Update tests.

* Update changelog.

* Bump pennylane version.

* Bump pennylane version.

* Auto update version

* Reimplement expval functors with macros.

* Auto update version

* Reimplement N-wires unitary gate functors with macros.

* Rename macros.

* Update CHANGELOG.md

* Auto update version

* trigger CI

* trigger CI

* Bump Kokkos to 4.1.00 in CI.

* Revert kokkos ver.

* Add tests for macroed expval functors.

* Remove redundant black lines.

* Add tests for macroed gate functors.

* Use matrix interface to get expval of HermitianObs in LKokkos.

* Fix few 1U bit shifts.

* Cover kokkos_args error.

* Define DoubleLoopRank in KokkosSV.

* Update changelog.

* Auto update version

* Fix codefactor error.

* Update pennylane_lightning/core/src/simulators/lightning_kokkos/gates/tests/Test_StateVectorKokkos_Param.cpp


* Auto update version

* Define i000 vars more explicitly outside the macros.

* trigger CI

* Auto update version

* trigger CI

* trigger CI

---------

+++ pennylane_lightning/core/src/simulators/lightning_kokkos/measurements/tests/Test_StateVectorKokkos_Measure.cpp
@@ -313,7 +313,8 @@ TEST_CASE("Test tensor transposition", "[Measure]") {
     SECTION("Looping over different wire configurations:") {
         for (auto &term : input) {
             // Defining a tensor to be transposed.
-            std::vector<size_t> indices(1U << term.first.size());
+            std::vector<size_t> indices(static_cast<std::size_t>(1U)
+                                        << term.first.size());
             std::iota(indices.begin(), indices.end(), 0);
 
             std::vector<size_t> results(indices.size());
